move some catalog tests
